Refactor Control to TaskList Storage now controls all Saving and Loading from file Ui now controls all UI interactions
To improve OOP
