+++ src/main/java/CreateFile.java
@@ -1,5 +1,4 @@
 import java.io.File;
-import java.io.FileReader;
 import java.io.IOException;
 
 public class CreateFile {
